code for traffic signal
Traffic light control systems are widely used to monitor and control 
the flow of automobiles through the junction of many roads. We aim to realize 
smooth motion of cars in the transportation routes. However, the 
synchronization of multiple traffic light systems at adjacent intersections is a 
complicated problem where various parameters are involved. Conventional 
systems do not handle variable flows approaching the junctions. In addition, 
the mutual interference between adjacent traffic light systems, the disparity 
of cars flow with time, the accidents, the passage of emergency vehicles, and 
the pedestrian crossing are not implemented in the existing traffic system. 
This leads to traffic jam and congestion.In this paper, we propose a method 
where a RF transmitter is placed in the vehicle and a receiver is placed at the 
traffic signals so that when the vehicle approaches the traffic light will respond 
and clears the traffic for the vehicle. The prototype of this project is using the 
frequency of 433 MHz and function with the sequence mode of traffic light 
when emergency vehicles passing by an intersection and changing the 
sequence back to the normal sequence before the emergency mode was 
triggered. In future, this prototype system can be improved by controlling the 
real traffic situation, in fact improving present traffic light system technology.